Move FFI to sub-crates
